% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac_epoch}
\alias{airac_epoch}
\title{AIRAC epoch as from ICAO Doc 8126}
\usage{
airac_epoch()
}
\value{
the AIRAC epoch according to ICAO scheme
}
\description{
AIRAC epoch is defined in  section 2.6.2 b) of
}
