% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac_interval}
\alias{airac_interval}
\title{The interval od dates for an (ICAO) AIRAC}
\usage{
airac_interval(airac)
}
\arguments{
\item{airac}{the ICAO AIRAC, i.e. "1603" for the 3rd AIRAC of 2016}
}
\value{
an interval of (28) days coverad by the AIRAC `airac`
}
\description{
The interval od dates for an (ICAO) AIRAC
}
\examples{
airac_interval("1603")

}
\seealso{
Other AIRAC functions: \code{\link{airac_year_epoch}},
  \code{\link{airac}}, \code{\link{cfmu_airac_interval}},
  \code{\link{cfmu_airac}}
}
\concept{AIRAC functions}
