% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac_year_epoch}
\alias{airac_year_epoch}
\title{Return the effective date of the first AIRAC for year \code{year}}
\usage{
airac_year_epoch(year)
}
\arguments{
\item{year}{a year (as integer)}
}
\value{
the WEF date for the first AIRAC of the year \code{year}
}
\description{
Return the effective date of the first AIRAC for year \code{year}
}
\examples{
airac_year_epoch(2018)
}
\seealso{
Other airac: 
\code{\link{airac}()},
\code{\link{airac_epoch}()},
\code{\link{airac_interval}()},
\code{\link{cfmu_airac}()},
\code{\link{cfmu_airac_epoch}()},
\code{\link{cfmu_airac_epoch_number}()},
\code{\link{cfmu_airac_interval}()}
}
\concept{airac}
