% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn.R
\name{arrivals_osn}
\alias{arrivals_osn}
\title{Get arrival at airport from Opensky Network}
\usage{
arrivals_osn(apt, wef, til = NULL)
}
\arguments{
\item{apt}{ICAO airport ID}

\item{wef}{starting from}

\item{til}{until time [default wef + 1 day]}
}
\description{
See \url{https://opensky-network.org/apidoc/index.html} for
details and limitations.
}
\examples{
\dontrun{
fligts_eddf <- arrivals_osn("EDDF", "2019-07-11")
}
}
