% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{bbox_at_distance}
\alias{bbox_at_distance}
\title{Return an axis-aligned bounding box delimiting a circle at distance `d`}
\usage{
bbox_at_distance(geo, d, ...)
}
\arguments{
\item{geo}{a geographical position [lon, lat]}

\item{d}{a distance in Nautical Miles}

\item{...}{other params}
}
\value{
an `st_bbox` object
}
\description{
\lifecycle{experimental}

`bbox_at_distance` returns an `st_bbox` object representing the extent of
an axis-aligned bounding box containing the (polygonal approximation of a)
circonference at distance `d` from the location, `geo`.

 WARNING: current implementation relies on `polygon_at_distance` which is
 not robust to cope with circles containing the Poles or crossing the date line.
}
\examples{
\dontrun{
fra <- c(8.570556, 50.03333) # Frankfurt Airport (longitude, latitude)
bbox_at_distance(fra, 40)
}
}
