% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{bounding_box}
\alias{bounding_box}
\title{Calculate Bounding Box}
\usage{
bounding_box(lat, lon, dist, in.miles = TRUE)
}
\arguments{
\item{lat}{The latitude of the center point.}

\item{lon}{The longitude of the center point.}

\item{dist}{The distance from the center point.}

\item{in.miles}{logical.  If \code{TRUE} uses miles as the units of
\code{dist}.  If \code{FALSE} uses kilometers.}
}
\value{
Returns a matrix with max/min latitude/longitude values.
}
\description{
Calculate a bounding box for a center point given a set of coordinates.
}
\examples{
bounding_box(38.8977, 77.0366, 1)
}
\references{
\url{http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates}
}
\keyword{bounding_box,}
\keyword{coordinates}
