% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{cfmu_airac}
\alias{cfmu_airac}
\title{return the CFMU AIRAC cycle number for a date}
\usage{
cfmu_airac(date)
}
\arguments{
\item{date}{a date}
}
\value{
a number indicating the AIRAC cycle according to CFMU scheme
}
\description{
return the CFMU AIRAC cycle number for a date
}
\examples{
cfmu_airac('2019-01-30')

}
\seealso{
Other AIRAC functions: \code{\link{airac_interval}},
  \code{\link{airac_year_epoch}}, \code{\link{airac}},
  \code{\link{cfmu_airac_interval}}
}
\concept{AIRAC functions}
