% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{cfmu_airac_epoch}
\alias{cfmu_airac_epoch}
\title{The CFMU epoch date for AIRAC numbering scheme}
\usage{
cfmu_airac_epoch()
}
\value{
the epoch date for CFMU AIRAC numbering scheme
}
\description{
The CFMU epoch date for AIRAC numbering scheme
}
