% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{cfmu_airac_epoch_number}
\alias{cfmu_airac_epoch_number}
\title{The CFMU epoch number for AIRAC numbering scheme}
\usage{
cfmu_airac_epoch_number()
}
\value{
the epoch number for CFMU AIRAC numbering scheme
}
\description{
The CFMU epoch number for AIRAC numbering scheme
}
