% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{cfmu_airac_interval}
\alias{cfmu_airac_interval}
\title{The interval of dates for a CFMU AIRAC number}
\usage{
cfmu_airac_interval(cfmu_airac)
}
\arguments{
\item{cfmu_airac}{a CFMU AIRAC number}
}
\value{
an interval of dates (28 days)
}
\description{
The interval of dates for a CFMU AIRAC number
}
\examples{
cfmu_airac_interval(441)

}
\seealso{
Other airac: 
\code{\link{airac_interval}()},
\code{\link{airac_year_epoch}()},
\code{\link{airac}()},
\code{\link{cfmu_airac}()}
}
\concept{airac}
