% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn-impala.R
\name{connect_osn}
\alias{connect_osn}
\title{Create an ssh session to OpenSky Network’s Impala shell.}
\usage{
connect_osn(usr, passwd = askpass, verbose = FALSE)
}
\arguments{
\item{usr}{user account}

\item{passwd}{either a string or a callback function for password prompt}

\item{verbose}{either TRUE/FALSE or a value between 0 and 4 indicating log level:
0: no logging, 1: only warnings, 2: protocol, 3: packets or 4: full stack trace.}
}
\value{
an SSH session
}
\description{
Create an ssh session to OpenSky Network’s Impala shell.
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
}
}
