% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cpr_plot_horizontal}
\alias{cpr_plot_horizontal}
\title{Plot CPR trajectories on a map}
\usage{
cpr_plot_horizontal(cprs, bb = NULL)
}
\arguments{
\item{cprs}{a dataframe of CPR position reports}

\item{bb}{an optional bounding box (a vector of left, bottom, right, top).
If NULL, the default, it will be calculated from the data.}
}
\value{
a ggplot2 plot
}
\description{
Plot CPR trajectories on a map
}
\examples{
\dontrun{
# plot CPR tryjectories
europe <- c(left = 5, bottom = 35, right = 30, top = 52)
cpr_plot_horizontal(poss, europe)
}
}
