% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-track.R
\name{cross_track_intersection}
\alias{cross_track_intersection}
\title{Find the cross track intersection between a point and an arc}
\usage{
cross_track_intersection(B, A1, A2)
}
\arguments{
\item{B}{Point from where to calculate the cross track intersection}

\item{A1}{First point of the arc}

\item{A2}{Second point of the arc}
}
\value{
A geographic coordinate representing the intersection point
}
\description{
Given an arc represented by two points A1 and A2, return the cross track
intersection for point B.
This is a helper function incorporating the code from example 11 of the
nvctr package
}
\examples{
# from example 11 of nvctr's vignette
cross_track_intersection(c(-1, -1), c(0, 3), c(0, 10))

}
