% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative_distance}
\alias{cumulative_distance}
\title{Calculate the cumulative distance (in km).}
\usage{
cumulative_distance(positions)
}
\arguments{
\item{positions}{the position reports for the flight}
}
\value{
a dataframe with a new \code{cumulative_distance} column (in km)
}
\description{
The calculated distance also for position reports on the ground;
the algorithm used is the one of \code{\link[geosphere]{distCosine}}.
}
\examples{
\dontrun{
cumulative_distance(poss)
# calculate only the flown distance
# NOTE: there is a gap in FR24 data between last ground position (if any)
#       and first in air one (typically at ~1500 ft).
#       The following code would certainly work better
#       would poss contain a position report for a very low altitude.
cumulative_distance(poss \%>\% filter(altitude > 0))
}
}
\seealso{
Other analysis: 
\code{\link{cumulative_time}()},
\code{\link{extract_segment}()},
\code{\link{filter_outlier}()},
\code{\link{resample}()},
\code{\link{smooth_positions}()}
}
\concept{analysis}
