% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative_time}
\alias{cumulative_time}
\title{Calculate the cumulative time (in s).}
\usage{
cumulative_time(positions, offblock = NULL, inblock = NULL)
}
\arguments{
\item{positions}{the position reports for the flight}

\item{offblock}{optional, the timestamp for offblock if
different from the first timestamp in the dataset.
It is either an \code{int}eger (UNIX timestamp) or a
date-time value}

\item{inblock}{optional, the timestamp for inblock if
different from the last timestamp in the dataset}
}
\value{
a dataframe with a new \code{cumulative_time} column (in min)
}
\description{
Calculate the cumulative time (in s).
}
\examples{
\dontrun{
cumulative_time(poss)
# calculate only the flown distance
# NOTE: there is a gap in FR24 data between last ground position (if any)
#       and first in air one (typically at ~1500 ft).
#       The following code would certainly work better
#       would poss contain a position report for a very low altitude.
cumulative_time(poss \%>\% filter(altitude > 0))
}
}
\seealso{
Other analysis: 
\code{\link{cumulative_distance}()},
\code{\link{extract_segment}()},
\code{\link{filter_outlier}()},
\code{\link{smooth_positions}()}
}
\concept{analysis}
