% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{ddm2dd}
\alias{ddm2dd}
\title{Convert Degree decimal Minute (DdM)) format to Decimal Degrees (DD).}
\usage{
ddm2dd(value)
}
\arguments{
\item{value}{a string representation of an angle (i.e. longitude) in
Degree decimal Minute (DdM)), i.e. N51°28.65'}
}
\value{
a decimal (degrees) value
}
\description{
Convert Degree decimal Minute (DdM)) format to Decimal Degrees (DD).
}
\examples{
ddm2dd("N51° 28.65'")

}
