% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn-impala.R
\name{departures_impala_osn}
\alias{departures_impala_osn}
\title{Get departures from airport}
\usage{
departures_impala_osn(session, apt, wef, til = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{apt}{ICAO ID of airport, i.e. "EDDF" for Frankfurt}

\item{wef}{Start of period of interest}

\item{til}{End of period of interest}
}
\value{
data frame of flight and track data (see OSN docs about
  \href{https://opensky-network.org/apidoc/rest.html#departures-by-airport}{Departures
   by Airport}).
}
\description{
Get departures from airport
}
\examples{
\dontrun{
depurtures_impala_osn(session, "EDDF", "2019-04-22 00:00:00", til=NULL)
}
}
