% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{dms2dd}
\alias{dms2dd}
\title{Convert  Degree Minute Seconds (DMS) format to Decimal Degrees (DD)}
\usage{
dms2dd(value)
}
\arguments{
\item{value}{a string representation of an angle (i.e. longitude) in
             Degree Minute Second (DMS)), i.e. "51° 43' 34.00\" N"

             \strong{Note} the escaping of double quote.}
}
\value{
a decimal degrees value
}
\description{
Convert  Degree Minute Seconds (DMS) format to Decimal Degrees (DD)
}
\examples{
ddm2dd("51° 43' 34.00\u0022 N")

}
\seealso{
Other coordinates: 
\code{\link{ddm2dd}()},
\code{\link{parse_lat_icao}()},
\code{\link{parse_lon_icao}()}
}
\concept{coordinates}
