% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_apt}
\alias{egll_apt}
\title{EGLL data.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 7 columns.}
\source{
UK AIP
}
\usage{
egll_apt
}
\description{
EGLL data.
}
\keyword{datasets}
