% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_flights}
\alias{egll_flights}
\title{Sample Reference Trajectories' flight data at EGLL.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45 rows and 9 columns.
}
\source{
Reference Trajectories
}
\usage{
egll_flights
}
\description{
Sample Reference Trajectories' flight data at EGLL.
}
\keyword{datasets}
