% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_flights_fr24}
\alias{egll_flights_fr24}
\title{Sample FlightRadar24's archived feed ADS-B flight data for EGLL.}
\format{
A data frame with xxyy rows and 11 variables:
\describe{
  \item{address}{mode-S address in hexadecimal, as supplied by the
   Mode-S or ADS-B transponder}
  \item{adep}{ICAO code for scheduled departure airport}
  \item{ades}{ICAO code for scheduled arrival airport}
  \item{callsign}{e.g. CSA190, as supplied by the Mode-S or
   ADS-B transponder}
  \item{flight_id}{unique identifier for the flight leg (integer)}
  \item{reg}{unique aircraft identifier (text)}
  \item{model}{aircraft type (e.g. B733, as per ICAO classification?),
  `GLID` for gliders,
  `GROUND` for ground vehicles}
  \item{flight}{flight number, e.g. OK190}
  \item{start_time}{date of the flight (what about the ones crssing midnight?)}
}
}
\source{
FlightRadar24
}
\usage{
egll_flights_fr24
}
\description{
A dataset containing FlightRadar24's archived feed ADS-B flight information
for XX flights.
There are xx flights on d1 and yy on yy.
}
\keyword{datasets}
