% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_positions}
\alias{egll_positions}
\title{Sample Reference Trajectories' position reports data at EGLL.}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18574 rows and 10 columns.
}
\source{
Reference Trajectories
}
\usage{
egll_positions
}
\description{
Sample Reference Trajectories' position reports data at EGLL.
}
\keyword{datasets}
