% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_positions_fr24}
\alias{egll_positions_fr24}
\title{Sample FlightRadar24's archived feed ADS-B position reports data at EGLL.}
\format{
A data frame with zzzz rows and 19 variables.
The variables are a joint of flight information from `flts` with
the following ones:
\describe{
  \item{flight_id}{unique identifier for the flight leg (integer)}
  \item{altitude}{altitude in feet (integer), as supplied by the Mode-S or
   ADS-B transponder}
  \item{latitude}{latitude in decimal degrees, data supplied by
   ADS-B but calculated for mode-S transponders and tracked using
   multi-lateration (MLAT)}
  \item{longitude}{longitude in decimal degrees, data supplied by
   ADS-B but calculated for mode-S transponders and tracked using
   multi-lateration (MLAT)}
  \item{speed}{ground speed in knots (integer), data supplied by
   ADS-B but calculated for mode-S transponders and tracked using
   multi-lateration (MLAT)}
  \item{event_time}{datetime for the position report}
  \item{on_ground}{1 = on ground, 0 = airborne}
  \item{vert_speed}{vertical speed}
}
}
\source{
FlightRadar24
}
\usage{
egll_positions_fr24
}
\description{
A dataset containing FlightRadar24's archived ADS-B position reports
(combined with flight information of `flts`) for xx flights at EGLL.
There are xx flights on d1 and yy on d2.
}
\keyword{datasets}
