% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egll_rw_sf}
\alias{egll_rw_sf}
\title{EGLL runways.}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.}
\source{
UK AIP
}
\usage{
egll_rw_sf
}
\description{
EGLL runways.
}
\keyword{datasets}
