% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{export_allft_so6}
\alias{export_allft_so6}
\title{Export \code{ALL_FT+}-based trajectories to \code{SO6} format}
\usage{
export_allft_so6(wef, til, model = "CTFM", ...)
}
\arguments{
\item{wef}{(UTC) timestamp of LOBT With Effect From (included).
Liberal format, i.e. "2019-07-14", "2019-07-14 10:21"
"2019-07-14T10:21:23Z"}

\item{til}{(UTC) timestamp of LOBT TILl (excluded).}

\item{model}{the trajectory model, one of FTFM, RTFM, CTFM, CPF}

\item{...}{extra arguments passed to \link{export_model_trajectory}.}
}
\value{
a dataframe of trajectory segments in SO6 format, see \code{\link{read_so6}}
for a description of the SO6 format.
}
\description{
Extract \code{ALL_FT+}-based trajectories from PRISME database and convert to
\code{SO6} format.
}
\details{
You need to store your credentials to access the PRU tables in
the following environment variables:
\itemize{
\item \code{PRU_TEST_USR} for the user id
\item \code{PRU_TEST_PWD} for the password
\item \code{PRU_TEST_DBNAME} for the database name
}
}
\examples{
\dontrun{
# BEWARE: this can take some long-ish time
export_allft_so6("2010-06-16", "2010-06-17", model = "FTFM")

# reduce the time scope to get the data quicker (and smaller)
export_allft_so6("2010-06-16 10:00", "2010-06-16T11:00:11")
}
}
\seealso{
Other read/export: 
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{generate_so6}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
