% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{export_allft_trajectory}
\alias{export_allft_trajectory}
\title{Export ALL_FT+-based NM trajectories}
\usage{
export_allft_trajectory(wef, til)
}
\arguments{
\item{wef}{(UTC) timestamp of With Effect From (included)}

\item{til}{(UTC) timestamp of TILl instant (excluded)}
}
\value{
a dataframe of trajectory points.
}
\description{
Extract "ALL_FT+"-based trajectories from PRISME database
}
\details{
You need to store your credentials to access the PRU related tables in
the following environment variables:
\itemize{
  \item \code{PRU_TEST_USR} for the user id
  \item \code{PRU_TEST_PWD} for the password
  \item \code{PRU_TEST_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_allft_trajectory("2010-06-16", "2010-06-17")
export_allft_trajectory("2010-06-16 10:00", "2010-06-16T11:00:00Z")
}
}
\seealso{
Other read/export functions: \code{\link{export_allft_so6}},
  \code{\link{export_event_so6}},
  \code{\link{export_event_trajectory}},
  \code{\link{export_flights_at_airport_fr24}},
  \code{\link{export_flights_fr24}},
  \code{\link{export_positions_at_airport_fr24}},
  \code{\link{export_positions_fr24}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_cpr}}, \code{\link{read_flights_fr24}},
  \code{\link{read_positions_fr24}}, \code{\link{read_so6}}
}
\concept{read/export functions}
