% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_apds}
\alias{export_apds}
\title{Export APDS (airport) data}
\usage{
export_apds(wef, til)
}
\arguments{
\item{wef}{date of With Effect From (included)}

\item{til}{date of TILl instant (excluded)}
}
\value{
a dataframe of airport reported movements
}
\description{
Extract APDS data from PRISME database.
**Note**: there are 129 columns...
}
\details{
You need to store your credentials to access the ATMAP tables in
the following environment variables:
\itemize{
  \item \code{PRU_ATMAP_USR} for the user id
  \item \code{PRU_ATMAP_PWD} for the password
  \item \code{PRU_ATMAP_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_apds("2019-04-10", "2019-04-11")
}
}
