% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{export_event_so6}
\alias{export_event_so6}
\title{Extract "Event"-based trajectories from CPLX tables (in PRISME) to SO6 format}
\usage{
export_event_so6(wef, til)
}
\arguments{
\item{wef}{the With-EFfect date from when flights are selected.}

\item{til}{the unTIL date till when flights are selected.}
}
\value{
a dataframe of trajectory segments in SO6 format.
}
\description{
You need to store your credentials to access the CPLX tables in
the following environment variables:
\itemize{
  \item \code{PRU_CPLX_USR} for the user id
  \item \code{PRU_CPLX_PWD} for the password
  \item \code{PRU_CPLX_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_event_so6("2010-06-16", "2010-06-17")
}
}
