% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{export_flights_at_airport_fr24}
\alias{export_flights_at_airport_fr24}
\title{Export FlightRadar24 flight movements at an airport in an interval of time}
\usage{
export_flights_at_airport_fr24(wef, til, apt, flow = "ALL")
}
\arguments{
\item{wef}{(UTC) timestamp of With Effect From (included)}

\item{til}{(UTC) timestamp of TILl instant (excluded)}

\item{apt}{IATA code of an airport, i.g. PSA for Pisa "Galileo Galilei"}

\item{flow}{the flow of flights: "ARR" for arrivals, "DEP" for departures,
"ALL" for both (default "ALL")}
}
\value{
a dataframe of flights
}
\description{
Extract  FlightRadar24 flight movements
in the specified period at an airport
}
\details{
You need to store your credentials to access the FR24 tables in
the following environment variables:
\itemize{
  \item \code{PRU_FR24_USR} for the user id
  \item \code{PRU_FR24_PWD} for the password
  \item \code{PRU_FR24_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_flights_at_airport_fr24("2017-09-01T00:00:00", "2017-09-02T00:00:00",
            "SVG",
            flow = "ARR")

# only 2 hours interval
export_flights_at_airport_fr24("2017-09-01 10", "2017-09-02T11:00:00",
            "SVG",
            flow = "ARR")
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{generate_so6}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
