% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{export_flights_fr24}
\alias{export_flights_fr24}
\title{Export FlightRadar24 flight movements for an interval of time}
\usage{
export_flights_fr24(wef, til)
}
\arguments{
\item{wef}{(UTC) timestamp of With Effect From (included)}

\item{til}{(UTC) timestamp of TILl instant (excluded)}
}
\value{
a dataframe of flights
}
\description{
Extract  FlightRadar24 flight movements
in the specified period
}
\details{
You need to store your credentials to access the FR24 tables in
the following environment variables:
\itemize{
  \item \code{PRU_FR24_USR} for the user id
  \item \code{PRU_FR24_PWD} for the password
  \item \code{PRU_FR24_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_flights_fr24("2017-09-01T10:30:00", "2017-09-01 11")
}
}
\seealso{
Other read/export functions: \code{\link{export_allft_so6}},
  \code{\link{export_allft_trajectory}},
  \code{\link{export_event_so6}},
  \code{\link{export_event_trajectory}},
  \code{\link{export_flights_at_airport_fr24}},
  \code{\link{export_positions_at_airport_fr24}},
  \code{\link{export_positions_fr24}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_cpr}}, \code{\link{read_flights_fr24}},
  \code{\link{read_positions_fr24}}, \code{\link{read_so6}}
}
\concept{read/export functions}
