% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_hourly_adsb}
\alias{export_hourly_adsb}
\title{Export list of hourly-binned flights and aircraft info}
\usage{
export_hourly_adsb(wef, til, model = "CTFM", bbox = NULL)
}
\arguments{
\item{wef}{(UTC) timestamp of LOBT With Effect From (included).
Liberal format, i.e. "2019-07-14", "2019-07-14 10:21"
"2019-07-14T10:21:23Z"}

\item{til}{(UTC) timestamp of LOBT TILl instant (excluded)}

\item{model}{the model of the profile: one of "FTFM", "CTFM", "CPF".
[default: "CFTM"]}

\item{bbox}{(Optional) axis aligned bounding box
(xmin, ymin, xmax, ymax)}
}
\value{
a dataframe with trajectory data
}
\description{
Extract hourly-binned flights from PRISME database together with aircraft information such
as registration, aircraft operator and ICAO 24-bit address
}
\details{
You need to store your credentials to access the PRU tables in
the following environment variables:
\itemize{
  \item \code{PRU_DEV_USR} for the user id
  \item \code{PRU_DEV_PWD} for the password
  \item \code{PRU_DEV_DBNAME} for the database name
}
}
\examples{
\dontrun{
# export all hourly CTFM flights within a bounding box 40 NM around EDDF
bb <- c(xmin = 7.536746, xmax = 9.604390, ymin = 49.36732, ymax = 50.69920)
export_hourly_adsb("2019-01-01 00:00", "2019-01-02 00:00", bbox = bb)
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_flights_csvs}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{generate_so6}()},
\code{\link{import_flights_csvs}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
