% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_movements}
\alias{export_movements}
\title{Export the flight list of movements at an airport}
\usage{
export_movements(wef, til, apt)
}
\arguments{
\item{wef}{(UTC) timestamp of LOBT With Effect From (included).
Liberal format, i.e. "2019-07-14", "2019-07-14 10:21"
"2019-07-14T10:21:23Z"}

\item{til}{(UTC) timestamp of LOBT TILl instant (excluded)}

\item{apt}{ICAO code of the airport, i.e. "EDDF"}
}
\value{
a data frame of flight list
}
\description{
Export the flight list of movements at an airport
}
\examples{
\dontrun{
export_movements("2020-01-20", "2020-01-21", "EDDF")
}
}
