% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{export_positions_fr24}
\alias{export_positions_fr24}
\title{Extract FR24 flights list for a time interval at an airport}
\usage{
export_positions_fr24(wef, til)
}
\arguments{
\item{wef}{timestamp of With Effect From (included)}

\item{til}{timestamp of TILl instant (excluded)}
}
\value{
a dataframe of position reports
}
\description{
You need to store your credentials to access the FR24 tables in
the following environment variables:
\itemize{
  \item \code{PRU_FR24_USR} for the user id
  \item \code{PRU_FR24_PWD} for the password
  \item \code{PRU_FR24_DBNAME} for the database name
}
}
\examples{
\dontrun{
library(ROracle)
export_flights_at_airport_fr24("2017-09-01T00:00:00Z", "2017-09-02T00:00:00Z",
            "SVG",
            flow = "ARR")
}
}
