% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfe.R
\name{extract_segment}
\alias{extract_segment}
\title{Extract level segment from a list of position reports.}
\usage{
extract_segment(prs, vert_limit = 5)
}
\arguments{
\item{prs}{dataset containing position reports.
Two variables are used: \code{timestamp} and \code{altitude} (feet)}

\item{vert_limit}{the maximum absolute vertical rate for a segment to be
considered level (default = 5 feet/s)}
}
\value{
a dataset of level segments. A level segment is defined by
        its beginning and end 4D position.
}
\description{
It is assumed that the dataset is grouped by flight id
and that position reports are sorted by ascending timestamp.
}
\examples{
\dontrun{
cumulative_time(poss)
f <- cprs \%>\%
 filter(callsign == "RYR40WJ") \%>\%
 arrange(timestamp_track) \%>\%
 rename(timestamp = timestamp_track) \%>\%
 mutate(altitude = flight_level * 100)
 extract_segment(f)
}
}
\seealso{
Other analysis: 
\code{\link{cumulative_distance}()},
\code{\link{cumulative_time}()},
\code{\link{filter_outlier}()},
\code{\link{smooth_positions}()}
}
\concept{analysis}
