% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{filter_positions_at_range}
\alias{filter_positions_at_range}
\title{Retain only positions within a range from a location.}
\usage{
filter_positions_at_range(
  df,
  geo,
  dm,
  dM,
  lon,
  lat,
  .exclude = FALSE,
  .keep = FALSE
)
}
\arguments{
\item{df}{a (trajectory) data frame}

\item{geo}{a geographical location in lon/lat (WGS84)}

\item{dm}{a distance in Nautical Miles}

\item{dM}{a distance in Nautical Miles}

\item{lon}{the column for longitude in \code{df}}

\item{lat}{the column for latitude in \code{df}}

\item{.exclude}{exclude the point in the (\code{dm}, \code{dM}) [default is FALSE]}

\item{.keep}{keep the calculated distance (in Nautical Miles)
in the \code{.distance} column [default is FALSE]}
}
\value{
a subset of \code{df}
}
\description{
The points whose distance, \code{.distance}, satisfies
\deqn{dm <= .distance < dM}
are kept (\code{.exclude = FALSE}) or excluded (\code{.exclude = TRUE})
}
\examples{
\dontrun{
fra <- c(8.570556, 50.03333) # Frankfurt Airport (longitude, latitude)

# keep the points 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 0, 40, longitude, latitude)
# keep the points from 10 to 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 10, 40, longitude, latitude)
# exclude the points from 10 to 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 10, 40, longitude, latitude, .exclude = TRUE)
# keep the points further away of 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 0, 40, longitude, latitude, .exclude = TRUE)
}
}
\seealso{
Other spatial: 
\code{\link{bbox_at_distance}()},
\code{\link{bounding_box}()},
\code{\link{crs_tc}()},
\code{\link{polygon_at_distance}()},
\code{\link{polygon_from_coords}()}
}
\concept{spatial}
