% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived.R
\name{fix_flights_csvs}
\alias{fix_flights_csvs}
\title{Fix a list flights' CSV files where \code{callsign}s contains ASCII NUL characters.}
\usage{
fix_flights_csvs(files, fixed_dir = getwd())
}
\arguments{
\item{files}{a list of the (full filenames) flights' files.}

\item{fixed_dir}{the directory where to save the fixed files
(defaults to current working directory).}
}
\value{
a list of newly created files where each NUL in \code{callsign}
        has been substituted by a SPACE character.
}
\description{
Fix a list flights' CSV files where \code{callsign}s contains ASCII NUL characters.
}
\examples{
\dontrun{
# fix all February 2017 flights files
flights_dir <- system.file("extdata", package = "trrrj")
flights_csvs <- dir(flights_dir, pattern = "201702.._flights\\\\.csv", full.names = TRUE)
flights_csvs <- fix_flights_csvs(flights_csvs)
}
}
\seealso{
\code{\link{fix_nulls_in_callsign}} for the new files naming convention.

Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{generate_so6}()},
\code{\link{import_flights_csvs}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
