% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{generate_so6}
\alias{generate_so6}
\title{Export trajectory profiles to SO6 format}
\usage{
generate_so6(trajectory)
}
\arguments{
\item{trajectory}{A data frame for point trajectories.}
}
\value{
A data frame for trajectories in SO6 format, see \code{\link{read_so6}}
        for a description of the SO6 format.
}
\description{
The data frame for point trajectories needs to have the following columns:

\tabular{llll}{
\strong{Name}            \tab \strong{Description}            \tab \strong{Type} \cr
\code{flight_id}         \tab Flight ID                       \tab int           \cr
\code{time_over}         \tab Time over point                 \tab datetime      \cr
\code{longitude}         \tab Longitude (decimal degrees)     \tab double        \cr
\code{latitude}          \tab Latitude (decimal degrees)      \tab double        \cr
\code{flight_level}      \tab Flight level                    \tab int           \cr
\code{point_id}          \tab Point ID or NO_POINT            \tab char          \cr
\code{air_route}         \tab Air route or NO_ROUTE           \tab char          \cr
\code{lobt}              \tab Last Off-block Time             \tab datetime      \cr
\code{seq_id}            \tab Positions's sequence number     \tab int           \cr
\code{callsign}          \tab Flight call sign                \tab char          \cr
\code{registration}      \tab Aircraft registration           \tab char          \cr
\code{model}             \tab Aircraft model                  \tab char          \cr
\code{aircraft_type}     \tab Aircraft ICAO type              \tab char          \cr
\code{aircraft_operator} \tab Aircraft operator               \tab char          \cr
\code{adep}              \tab Departing aerodrome (ICAO) ID   \tab char          \cr
\code{ades}              \tab Destination aerodrome (ICAO) ID \tab char
}
}
\examples{
\dontrun{
generate_so6(trj)
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_flights_csvs}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{import_flights_csvs}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
