% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived.R
\name{import_flights_csvs}
\alias{import_flights_csvs}
\title{Import various FR24 flights CSV files.}
\usage{
import_flights_csvs(files)
}
\arguments{
\item{files}{a list of FR24 flights CSV files}
}
\value{
a data frame of all imported flights.
        Use \code{attr(df, "problems")} for issues with the data read.
}
\description{
Import various FR24 flights CSV files.
}
\examples{
\dontrun{
# read all package February 2017 flights files
flights_dir <- system.file("extdata", package = "trrrj")
flights_csvs <- dir(flights_dir, pattern = "201702.._flights\\\\.csv", full.names = TRUE)
flts <- import_flights_csvs(flights_csvs)
pbs <- attr(flts, "problems")
pbs
# the problematic rows have truncated values after the embedded NUL,
# so filter them out ...
bad_rows <- pbs \%>\% pull(row)
flts \%>\% filter(!(row_number() \%in\% bad_rows))

# ...or use fix_flights_csvs()
}
}
\seealso{
\code{\link{read_flights_fr24}} for FR24 archive file naming convention.

\code{\link{fix_flights_csvs}} for reading files with embedded NUL's
         in callsign.

Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_flights_csvs}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{generate_so6}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
