% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{parse_lat_icao}
\alias{parse_lat_icao}
\title{Parse latitude in ICAO format}
\usage{
parse_lat_icao(lat)
}
\arguments{
\item{lat}{a string representation for latitude,
i.e. \code{554718N} or \code{554718.23N}}
}
\value{
a latitude in decimal degrees
}
\description{
The format is as described in ICAO Doc 8126 section 5.5.2
bullet d):
\preformatted{
  In the indication of the geographical coordinates of
  a location:
  - the latitude should be given first;
  - symbols for degrees, minutes or seconds should
    be omitted;
  - two digits should always be used in expressing
    values of less than 10 degrees of latitude; and
  - three digits should always be used in expressing
    values of less than 100 degrees of longitude.
  For example, 050735N 0652542W means five
  degrees, seven minutes and thirty-five seconds
  North, sixty-five degrees, twenty-five minutes and
  forty-two seconds West.
}
What is typical in AIP pubblications is the additional part
of decimal seconds, like \code{554718.23N}.
}
\details{
This format is used in AIP pubblications, for examples for runway threshold
coordinates, stand positions, etc.
This is used also in `ALL+` files from NM.
}
\examples{
parse_lat_icao("554718N")

}
\seealso{
Other coordinatetes conversion functions: \code{\link{ddm2dd}},
  \code{\link{dms2dd}}, \code{\link{parse_lon_icao}}
}
\concept{coordinatetes conversion functions}
