% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{parse_lon_icao}
\alias{parse_lon_icao}
\title{Parse longitude in ICAO format}
\usage{
parse_lon_icao(lon)
}
\arguments{
\item{lon}{a string representation for longitude, i.e. 0114901W or 0114901.23W}
}
\value{
decimal degress value
}
\description{
The format is as described in ICAO Doc 8126 section 5.5.2
bullet d):
\preformatted{
  In the indication of the geographical coordinates of
  a location:
  - the latitude should be given first;
  - symbols for degrees, minutes or seconds should
    be omitted;
  - two digits should always be used in expressing
    values of less than 10 degrees of latitude; and
  - three digits should always be used in expressing
    values of less than 100 degrees of longitude.
  For example, 050735N 0652542W means five
  degrees, seven minutes and thirty-five seconds
  North, sixty-five degrees, twenty-five minutes and
  forty-two seconds West.
}
What is typical in AIP pubblications is the additional part
of decimal seconds, like `554718.23N`.
}
\details{
This format is used in AIP pubblications, for examples for runway threshold
coordinates, stand positions, etc.
This is used also in `ALL+` files from NM.
}
\examples{
parse_lon_icao("0114901W")

}
\seealso{
Other coordinatetes conversion functions: 
\code{\link{ddm2dd}()},
\code{\link{dms2dd}()},
\code{\link{parse_lat_icao}()}
}
\concept{coordinatetes conversion functions}
