% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cpr_horizontal}
\alias{plot_cpr_horizontal}
\title{Plot CPR trajectories on a map}
\usage{
plot_cpr_horizontal(poss, bbox = NULL, ...)
}
\arguments{
\item{poss}{a dataframe of correlated position reports (CPRs) with (at least)
\code{callsign}, \code{timestamp_track} (a date-time), \code{flight_level} (in flight levels),
\code{longitude} (in decimal degrees) and \code{latitude} (in decimal degrees) columns}

\item{bbox}{a bounding box in the format
\code{c(lowerleftlon, lowerleftlat, upperrightlon, upperrightlat)}.}

\item{...}{Extra parameters to pass further on.}
}
\value{
a ggplot2 plot
}
\description{
Plot CPR trajectories on a map
}
\examples{
\dontrun{
# plot CPR tryjectories
europe <- c(left = 5, bottom = 35, right = 30, top = 52)
plot_cpr_horizontal(cprs, europe)
}
}
\seealso{
Other plot: 
\code{\link{plot_flight_horizontal}()},
\code{\link{plot_flight_vertical_distance}()},
\code{\link{plot_flight_vertical_time}()}
}
\concept{plot}
