% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flight_vertical_distance}
\alias{plot_flight_vertical_distance}
\title{Plot the vertical profile of the recorded positions of a flight from
lapsed distance perspective.}
\usage{
plot_flight_vertical_distance(poss)
}
\arguments{
\item{poss}{a dataframe of position reports with (at least)
`timestamp` (a date-time), `altitude` (in feet), `longitude`
(in decimal degrees) and `latitude` (in decimal degrees) and
`callsign` columns}
}
\value{
a \code{ggplot2} plot object.
}
\description{
Plot the vertical profile of the recorded positions of a flight from
lapsed distance perspective.
}
\examples{
\dontrun{
plot_flight_vertical_distance(poss)
}
}
