% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flight_vertical_time}
\alias{plot_flight_vertical_time}
\title{Plot the vertical profile of the recorded positions of a flight from
lapsed time perspective.}
\usage{
plot_flight_vertical_time(poss)
}
\arguments{
\item{poss}{a dataframe of position reports with (at least)
`timestamp` (a date-time) and `altitude` (in feet) columns}
}
\value{
a \code{ggplot2} plot object.
}
\description{
Plot the vertical profile of the recorded positions of a flight from
lapsed time perspective.
}
\examples{
\dontrun{
plot_flight_vertical_time(poss)
}
}
\seealso{
Other plot functions: 
\code{\link{plot_cpr_horizontal}()},
\code{\link{plot_flight_horizontal}()},
\code{\link{plot_flight_vertical_distance}()}
}
\concept{plot functions}
