% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{polygon_from_coords}
\alias{polygon_from_coords}
\title{Return a polygon from (WGS84) coordinates}
\usage{
polygon_from_coords(coords, dfMaxLength = units::set_units(1, "km"))
}
\arguments{
\item{coords}{coordinates defining a polygon}

\item{dfMaxLength}{maximum length of a line segment. If \code{x} has geographical coordinates (long/lat), \code{dfMaxLength} is either a numeric expressed in meter, or an object of class \code{units} with length units \code{rad} or \code{degree}; segmentation in the long/lat case takes place along the great circle, using \link[lwgeom:geod]{st_geod_segmentize}.}
}
\value{
an sf polygon
}
\description{
Return a polygon from (WGS84) coordinates
}
\examples{
\dontrun{
coords <- list(rbind(c(0, 0), c(1, 0), c(1, 1), c(0.5, 1.5), c(0, 1), c(0, 0)))
polygon_from_coords(coords)
}
}
\seealso{
Other spatial: 
\code{\link{bbox_at_distance}()},
\code{\link{bounding_box}()},
\code{\link{crs_tc}()},
\code{\link{filter_positions_at_range}()},
\code{\link{polygon_at_distance}()}
}
\concept{spatial}
