% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr.R
\name{read_cpr}
\alias{read_cpr}
\title{Read Network Manager's archived CPR file}
\usage{
read_cpr(file, delim = ";")
}
\arguments{
\item{file}{A (gzipped) file containing archived CPR's as received
received and processed by the Network Manager.
Filenames are something like
\code{1.201702061001tacop304ARCHIVED_OPLOG_ALL_CPR.gz}}

\item{delim}{the character delimiting the fields [default = ";"]}
}
\value{
A dataframe of correlated position reports (CPR's).
        The returned dataframe has the following variables:
        \tabular{lll}{
          Name \tab Description \tab Type\cr
          cpm_id \tab  \tab  \cr
          tact_id \tab ETFMS ID \tab string \cr
          timestamp_etfms \tab  \tab  \cr
          timestamp_track \tab  \tab  \cr
          block \tab  \tab  \cr
          record \tab  \tab  \cr
          entry_node_sac \tab  \tab  \cr
          entry_node_sic \tab  \tab  \cr
          callsign \tab Callsign of flight \tab string \cr
          adep_icao \tab ICAO code of Departure aerodrome \tab string \cr
          ades_icao \tab ICAO code of Destination aerodrome \tab string \cr
          eobt \tab Estimated Off-block Time \tab timestamp \cr
          longitude \tab Longitude (WGS84) \tab decimal degrees  \cr
          latitude \tab  Latitude (WGS84) \tab  decimal degrees \cr
          flight_level \tab  Flight level \tab  integer \cr
          track_service \tab  \tab  \cr
          ssr_code \tab  \tab  \cr
          track_speed \tab  \tab  \cr
          track_heading \tab  \tab  \cr
          climb_rate \tab  \tab  \cr
          track_vertical_mode \tab vertical attitude \tab factor  \cr
          ifps_id \tab  \tab  \cr
          aircraft_address \tab ICAO 24-bit address \tab string
        }
}
\description{
Read Network Manager's archived CPR file
}
\examples{
\dontrun{
# read CPR's of the 20 sample flights for 5th Feb 2017
cprs_dir <- system.file("extdata", package = "trrrj")
cprs_file <- paste0(cprs_dir, "/1.201702051001tacop304ARCHIVED_OPLOG_ALL_CPR.gz")
cprs05 <- read_cpr(cprs_file)
}
}
\seealso{
Other read/export functions: \code{\link{export_event_so6}},
  \code{\link{export_event_trajectory}},
  \code{\link{export_flights_at_airport_fr24}},
  \code{\link{export_flights_fr24}},
  \code{\link{export_positions_at_airport_fr24}},
  \code{\link{export_positions_fr24}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_flights_fr24}},
  \code{\link{read_positions_fr24}}, \code{\link{read_so6}}
}
\concept{read/export functions}
