% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{read_positions_fr24}
\alias{read_positions_fr24}
\title{Read FR24's archived positions information for a flight.}
\usage{
read_positions_fr24(path)
}
\arguments{
\item{path}{the filename of the \code{<yyyymmdd>_<flight_id>_positions.csv} file}
}
\value{
a data frame of the archived flight's positions for date \code{<yyyymmdd>}.
  Each position report has the following attributes (all mandatory, i.e. non empty,
  but \code{speed}, \code{squawk} and \code{vert_speed}):
  \itemize{
    \item \code{flight_id}   FR24 unique (within a day?) flight leg identifier
    \item \code{timestamp}   \code{snapshot_id} converted to a date/time
    \item \code{snapshot_id} time of sample in Unix epoch seconds,
          i.e. since 1st Jan 1970 00:00:00 UTC
    \item \code{altitude}    altitude (feet). Barometric relative to ISO 1013 hPa
          pressure in flight, always 0 on ground.
    \item \code{heading}     0-359 degrees
    \item \code{latitude}    latitude in decimal degrees
    \item \code{longitude}   longitude in decimal degrees
    \item \code{radar_id}    unique identifier of primary ground receiver
    \item \code{speed}       ground speed (knots)
    \item \code{squawk}      code broadcast by airplane, represented as four
          octal digits
    \item \code{vert_speed}  vertical speed (feet/min). Calculated from rate
          of altitude changes, 0 for level flight or empty if unavailable
    \item \code{on_ground}   1 for on ground, 0 for airborne.
          Calculated from low altitude and ground speed
  }
}
\description{
\code{read_positions_fr24} returns all position reports as archived by FR24
for a given \code{flight_id} flown on a specific day (see \code{path}
parameter for how the date and flight ID are encoded).
}
\details{
The positions listed are the ones known to FR24 and fitting the
the bounding box as contractually agreed.
}
\examples{
\dontrun{
# positions for flight 207522820 on 20170206 (6th Feb 2017)
poss_dir <- system.file("extdata", package = "trrrj")
poss_file <- paste0(poss_dir, "/20170206_positions/20170206_207522820.csv")
read_positions_fr24()
}
}
\seealso{
Other archive import functions: \code{\link{fix_flights_csvs}},
  \code{\link{fix_nulls_in_callsign}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_cpr}}, \code{\link{read_flights_fr24}}
}
\concept{archive import functions}
