% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{read_so6}
\alias{read_so6}
\title{Read SO6 file}
\usage{
read_so6(filename, delim = " ")
}
\arguments{
\item{filename}{the file containing SO6 trajectories}

\item{delim}{the field delimiter (default: " " [blank])}
}
\value{
a data frame where date and time are combined in a single
        datetime and longitude and latitude are in decimal degrees.
}
\description{
The SO6 file has 20 variables of roughly two types:
* Segment-specific data (latitude/longitude, time/data information)
* Flight meta data (flight identifier, callsign, aircraft type)
}
\examples{
\dontrun{
so6file <- system.file("extdata",
                       "TRAFFIC_20180630_reduced.so6",
                       package = "trrrj")
read_so6(so6file)
}
}
