% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{.data}
\alias{askpass}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{askpass}{\code{\link[askpass]{askpass}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{.data}}}
}}

