% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{resample a bunch of flight trajectories}
\usage{
resample(df, interval = 1L)
}
\arguments{
\item{df}{a data frame of flight trajectories, each identified by
its own \code{flight_id}}

\item{interval}{the amount of seconds to resample to}
}
\value{
a data frame of interpolated 4D positions at \code{interval} seconds
}
\description{
Each flight trajectory 4D positions are linearly interpolated at
\code{interval} seconds. First and last positions are retained.
}
\examples{
\dontrun{
poss |>
  resample(300L)
}
}
