% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_4d}
\alias{resample_4d}
\title{resample 4D a single flight}
\usage{
resample_4d(df, interval)
}
\arguments{
\item{df}{a data frame for a single flight trajectory}

\item{interval}{amount of seconds to resample to}
}
\value{
a data frame with interpolated 4D values at the requested \code{interval}
}
\description{
resample 4D a single flight
}
