% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn-impala.R
\name{state_vector_impala_osn}
\alias{state_vector_impala_osn}
\title{Get state vectors from OSN}
\usage{
state_vector_impala_osn(session, icao24, wef_time, til_time = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{icao24}{(Optional) Single or vector of ICAO24 ICAO 24-bit addresses}

\item{wef_time}{Start of period of interest}

\item{til_time}{End of period of interest}
}
\value{
a dataframe of state vectors
}
\description{
Get state vectors from OSN
}
\examples{
\dontrun{
state_vector_impala_osn(session, icao24 = c("3c6589", "3c6757"), wef_time = "2019-04-22 00:00:00", til_time = "2019-04-22 10:00:00")
}
}
