% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trrrj-package.R
\docType{package}
\name{trrrj-package}
\alias{trrrj}
\alias{trrrj-package}
\title{trrrj: Flight trajectories analysis}
\description{
Functions to a) read FlightRadar24's live feed JSON files, archived data; b) import CPR data from NM's archived files; c) Plot 2D trajectories, vertical profiles both in time and track distance; d) analyse trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trrrj.ansperformance.eu}
  \item \url{https://github.com/euctrl-pru/trrrj}
  \item Report bugs at \url{https://github.com/euctrl-pru/trrrj/issues}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
