% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{trrrj_example}
\alias{trrrj_example}
\title{Get path to trrrj example}
\usage{
trrrj_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
trrrj comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
trrrj_example()
trrrj_example("20170205_flights.csv")
}
