% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac}
\alias{airac}
\title{Return the ICAO IRAC id for a date}
\usage{
airac(date)
}
\arguments{
\item{date}{a date (or a ymd representation of it)}
}
\value{
the ICAO AIRCA id
}
\description{
Return the ICAO IRAC id for a date
}
\examples{
airac('2019-01-30')

}
\seealso{
Other airac: 
\code{\link{airac_epoch}()},
\code{\link{airac_interval}()},
\code{\link{airac_year_epoch}()},
\code{\link{cfmu_airac_interval}()},
\code{\link{cfmu_airac}()}
}
\concept{airac}
