% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac_epoch}
\alias{airac_epoch}
\title{AIRAC epoch as from ICAO Doc 8126}
\usage{
airac_epoch()
}
\value{
the AIRAC epoch according to ICAO scheme
}
\description{
AIRAC epoch is defined in  section 2.6.2 b) of
}
\examples{
airac_epoch()

}
\seealso{
Other airac: 
\code{\link{airac}()},
\code{\link{airac_interval}()},
\code{\link{airac_year_epoch}()},
\code{\link{cfmu_airac}()},
\code{\link{cfmu_airac_epoch}()},
\code{\link{cfmu_airac_epoch_number}()},
\code{\link{cfmu_airac_interval}()}
}
\concept{airac}
