% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{airac_year_epoch}
\alias{airac_year_epoch}
\title{Return the effective date of the first AIRAC for year `year`}
\usage{
airac_year_epoch(year)
}
\arguments{
\item{year}{a year (as integer)}
}
\value{
the WEF date for the first AIRAC of the year `year`
}
\description{
Return the effective date of the first AIRAC for year `year`
}
\examples{
airac_year_epoch(2018)
}
