% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn-impala.R
\name{arrivals_impala_osn}
\alias{arrivals_impala_osn}
\title{Get arrivals at airport}
\usage{
arrivals_impala_osn(session, apt, wef, til = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{apt}{ICAO ID of airport, i.e. "EDDF" for Frankfurt}

\item{wef}{Start of period of interest}

\item{til}{End of period of interest}
}
\value{
data frame of state vector data as for OSN docs.
}
\description{
Get arrivals at airport
}
\examples{
\dontrun{
arrivals_impala_osn(session, "EDDF", "2019-04-22 00:00:00", til=NULL)
}
}
