% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{bounding_box}
\alias{bounding_box}
\title{Calculate the coordinates of the axis-aligned bounding box}
\usage{
bounding_box(lat, lon, d)
}
\arguments{
\item{lat}{latitude of the center point  [decimal degrees].}

\item{lon}{longitude of the center point [decimal degrees].}

\item{d}{distance from the center point in Nautical Miles.}
}
\value{
Returns a matrix with max/min latitude/longitude values.
}
\description{
\lifecycle{experimental}

Calculate a bounding box for a center point given a set of coordinates.
}
\examples{
\dontrun{
bounding_box(38.8977, 77.0366, 1)
}
}
\references{
\url{http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates}
}
\seealso{
Other spatial: 
\code{\link{bbox_at_distance}()},
\code{\link{crs_tc}()},
\code{\link{filter_positions_at_range}()},
\code{\link{polygon_at_distance}()}
}
\concept{spatial}
\keyword{bounding_box,}
\keyword{coordinates}
