% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-airac.R
\name{cfmu_airac_epoch}
\alias{cfmu_airac_epoch}
\title{The CFMU epoch date for AIRAC numbering scheme}
\usage{
cfmu_airac_epoch()
}
\value{
the epoch date for CFMU AIRAC numbering scheme
}
\description{
The CFMU epoch date for AIRAC numbering scheme
}
\seealso{
Other airac: 
\code{\link{airac}()},
\code{\link{airac_epoch}()},
\code{\link{airac_interval}()},
\code{\link{airac_year_epoch}()},
\code{\link{cfmu_airac}()},
\code{\link{cfmu_airac_epoch_number}()},
\code{\link{cfmu_airac_interval}()}
}
\concept{airac}
\keyword{internal}
