% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{crs_tc}
\alias{crs_tc}
\title{PROJ4 CRS string used in Traffic Complexity}
\usage{
crs_tc()
}
\value{
the PROJ4 string for the map projection
}
\description{
Traffic Complexity score computation uses a custom Albers Equal Area projection.
}
\seealso{
\code{\link{parse_airspace_prisme}} for an example.

Other spatial: 
\code{\link{bbox_at_distance}()},
\code{\link{bounding_box}()},
\code{\link{filter_positions_at_range}()},
\code{\link{polygon_at_distance}()},
\code{\link{polygon_from_coords}()}
}
\concept{spatial}
