% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn-impala.R
\name{disconnect_osn}
\alias{disconnect_osn}
\title{Disconnect from OpenSky Network’s Impala shell.}
\usage{
disconnect_osn(session)
}
\arguments{
\item{session}{ssh connection created with \code{\link[ssh:ssh_connect]{ssh_connect()}}}
}
\value{
an SSH session
}
\description{
Disconnect from OpenSky Network’s Impala shell.
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
disconnect_osn(session)
}
}
