% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_apds}
\alias{export_apds}
\title{Export APDS (airport) data}
\usage{
export_apds(wef, til)
}
\arguments{
\item{wef}{date of With Effect From (included)}

\item{til}{date of TILl instant (excluded)}
}
\value{
a dataframe of airport reported movements
}
\description{
Extract APDS data from PRISME database.
}
\details{
You need your ATMAP DB credential stored in the \code{PRU_ATMAP_USR},
\code{PRU_ATMAP_PWD} and \code{PRU_ATMAP_DBNAME} environment
variables.
}
\examples{
\dontrun{
export_apds("2019-04-10", "2019-04-11")
}
}
