% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_apds}
\alias{export_apds}
\title{Export APDS (airport) data}
\usage{
export_apds(wef, til)
}
\arguments{
\item{wef}{date of With Effect From (included)}

\item{til}{date of TILl instant (excluded)}
}
\value{
a dataframe of airport reported movements
}
\description{
Extract APDS data from PRISME database.
**Note**: there are 129 columns...
}
\details{
You need to store your credentials to access the ATMAP tables in
the following environment variables:
\itemize{
  \item \code{PRU_ATMAP_USR} for the user id
  \item \code{PRU_ATMAP_PWD} for the password
  \item \code{PRU_ATMAP_DBNAME} for the database name
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_flights_csvs}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{generate_so6}()},
\code{\link{import_flights_csvs}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
