% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so6.R
\name{export_event_trajectory}
\alias{export_event_trajectory}
\title{Export event-based NM trajectories}
\usage{
export_event_trajectory(wef, til)
}
\arguments{
\item{wef}{the With-EFfect date from when flights are selected.}

\item{til}{the unTIL date till when flights are selected.}
}
\value{
a dataframe of trajectory points.
}
\description{
Extract "Event"-based trajectories from PRISME database
}
\details{
You need to store your credentials to access the CPLX related tables in
the following environment variables:
\itemize{
  \item \code{PRU_CPLX_USR} for the user id
  \item \code{PRU_CPLX_PWD} for the password
  \item \code{PRU_CPLX_DBNAME} for the database name
}
}
\examples{
\dontrun{
export_event_trajectory("2010-06-16", "2010-06-17")
export_event_trajectory("2010-06-16 10:00", "2010-06-16T11:00:00Z")
}
}
\seealso{
Other read/export functions: \code{\link{export_event_so6}},
  \code{\link{export_flights_at_airport_fr24}},
  \code{\link{export_flights_fr24}},
  \code{\link{export_positions_at_airport_fr24}},
  \code{\link{export_positions_fr24}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_cpr}}, \code{\link{read_flights_fr24}},
  \code{\link{read_positions_fr24}}, \code{\link{read_so6}}
}
\concept{read/export functions}
