% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_flight_info}
\alias{export_flight_info}
\title{Export the flight list}
\usage{
export_flight_info(wef, til, lobt_buffer = c(before = 28, after = 24))
}
\arguments{
\item{wef}{(UTC) timestamp of LOBT With Effect From (included).
Liberal format, i.e. "2019-07-14", "2019-07-14 10:21"
"2019-07-14T10:21:23Z"}

\item{til}{(UTC) timestamp of LOBT TILl instant (excluded)}

\item{lobt_buffer}{The number of hours before and after LOBT to query
[default before = 28, after = 24]. This is related to how
LOBT is stored in the underlying database table.}
}
\value{
a data frame of flight info
}
\description{
Export the flight list
}
\examples{
\dontrun{
export_flight_info("2020-01-20", "2020-01-21")
}
}
