% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_model_trajectory}
\alias{export_model_trajectory}
\title{Export point profile from NM trajectories}
\usage{
export_model_trajectory(wef, til, model = "CTFM")
}
\arguments{
\item{wef}{the day With-Wffect-From, i.e. "2019-07-14"}

\item{til}{the day unTIL, i.e. "2019-07-16" (not included)}

\item{model}{the model of the profile: one of "FTFM", "CTFM", "CPF".
[default: "CFTM"]}
}
\value{
a dataframe with trajectory data
}
\description{
Export point profile from NM trajectories
}
\examples{
\dontrun{
export_model_trajectory("2019-07-14", "2019-07-15")
}
}
