% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{export_model_trajectory}
\alias{export_model_trajectory}
\title{Export point profile from NM trajectories}
\usage{
export_model_trajectory(wef, til, model = "CTFM")
}
\arguments{
\item{wef}{(UTC) timestamp of LOBT With Effect From (included).
Liberal format, i.e. "2019-07-14", "2019-07-14 10:21"
"2019-07-14T10:21:23Z"}

\item{til}{(UTC) timestamp of LOBT TILl instant (excluded)}

\item{model}{the model of the profile: one of "FTFM", "CTFM", "CPF".
[default: "CFTM"]}
}
\value{
a dataframe with trajectory data
}
\description{
Extract NM point profile trajectories from PRISME database
}
\details{
You need to store your credentials to access the PRU tables in
the following environment variables:
\itemize{
  \item \code{PRU_DEV_USR} for the user id
  \item \code{PRU_DEV_PWD} for the password
  \item \code{PRU_DEV_DBNAME} for the database name
}
}
\examples{
\dontrun{
# export 1 day worth of NM (planned) trajectories
export_model_trajectory("2019-07-14", "2019-07-15", model = "FTFM")

# export 2 hours of NM (flown) trajectories
export_model_trajectory("2019-07-14 22:00", "2019-07-15")
}
}
