% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{filter_positions_at_range}
\alias{filter_positions_at_range}
\title{Retain only positions within a range from a location.}
\usage{
filter_positions_at_range(
  df,
  geo,
  dm,
  dM,
  lon,
  lat,
  .exclude = FALSE,
  .keep = FALSE
)
}
\arguments{
\item{df}{a (trajectory) data frame}

\item{geo}{a geographical location in lon/lat (WGS84)}

\item{dm}{a distance in Nautical Miles}

\item{dM}{a distance in Nautical Miles}

\item{lon}{the column for longitude in `df`}

\item{lat}{the column for latitude in `df`}

\item{.exclude}{exclude the point in the [`dm`, `dM`) [default is FALSE]}

\item{.keep}{keep the calculated distance (in Nautical Miles)
in the `.distance` column [default is FALSE]}
}
\value{
a subset of `df`
}
\description{
The points whose distance, `.distance`, satisfies
\deqn{dm <= .distance < dM}
are kept (`.exclude == FALSE`) or excluded (`.exclude == TRUE`)
}
\examples{
\dontrun{
fra <- c(8.570556, 50.03333) # Frankfurt Airport (longitude, latitude)

# keep the points 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 0, 40, longitude, latitude)
# keep the points from 10 to 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 10, 40, longitude, latitude)
# exclude the points from 10 to 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 10, 40, longitude, latitude, .exclude = TRUE)
# keep the points further away of 40 NM from FRA
poss \%>\% filter_positions_at_range(fra, 0, 40, longitude, latitude, .exclude = TRUE)
}
}
