% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived.R
\name{fix_nulls_in_callsign}
\alias{fix_nulls_in_callsign}
\title{Fix flights' CSV file where \code{callsign}s contains ASCII NUL characters.}
\usage{
fix_nulls_in_callsign(flights_csv, fixed_dir = getwd())
}
\arguments{
\item{flights_csv}{the full filename to the flights' file.}

\item{fixed_dir}{the directory where to save the fixed files
(defaults to current working directory).}
}
\value{
a the filename of a new file where each NUL in \code{callsign}
        has been substituted by a SPACE character.
        This typically happens for gliders (i.e. \code{equip == "GLID"}).
        The filename has a \code{_fixed} postfix to the original basename
        (without extension), i.e. \code{20170206_flights.csv} becomes
        \code{20170206_flights_fixed.csv}.
}
\description{
Fix flights' CSV file where \code{callsign}s contains ASCII NUL characters.
}
\examples{
\dontrun{
# fix 6th Feb 2017 flights file
flights_dir <- system.file("extdata", package = "trrrj")
flights_file <- paste0(flights_dir, "/20170206_flights.csv")
flights_csv <- fix_nulls_in_callsign(flights_file)
}
}
