% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airspace.R
\name{parse_airspace_prisme}
\alias{parse_airspace_prisme}
\title{Parse PRISME Airspace format}
\usage{
parse_airspace_prisme(lines)
}
\arguments{
\item{lines}{text lines of the PRISME airspace representation}
}
\value{
a tibble of airspaces (CRS = 4326)
}
\description{
\lifecycle{experimental}

\code{parse_airspace_prisme} reads the airspace PRISME format and returns a tibble
where each row describes an airspace.

Each airspace in PRISME format is described by:
\enumerate{
\item Airspace description (one row):
\itemize{
\item an airspace Unit code
\item a Elementary Sector code
\item minimum flight level
\item maximum flight level
\item the number of vertices of the polygon of the Elementary Sector
\item the coordinates of vertices
}
\item vertex coordinates (as many rows as the number of vertices)
}
}
\examples{
\dontrun{
es <- system.file("extdata", "airspace_prisme_462.prisme", package = "trrrj")
sectors <- readr::read_lines(es) \%>\%
  parse_airspace_prisme()
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{generate_so6}()},
\code{\link{read_cpr}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
