% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{polygon_at_distance}
\alias{polygon_at_distance}
\title{Generate the polygon at distance d from a geographical location}
\usage{
polygon_at_distance(geo, d, ...)
}
\arguments{
\item{geo}{a geographical location in lon/lat (WGS84)}

\item{d}{a distance in Nautical Miles}

\item{...}{other parameters, for example \code{nQuadSegs};
see also \code{\link[sf:st_buffer]{st_buffer}}}
}
\value{
a polygon.
}
\description{
\lifecycle{experimental}

\code{polygon_at_distance} returns a polygon approximating a circonference
at distance \code{d}, in Nautical Miles, from the location \code{geo}.
You can control how many points per quadrant will be used via the
\code{nQuadSegs} parameter (the default of 30 from \link[sf:st_buffer]{st_buffer} should
suffice for most of the needs.)

WARNING: this is not tested to work across the date line or for polygons
containing the poles or for polygons spanning more than half an emisphere.
}
\examples{
\dontrun{
fra <- c(8.570556, 50.03333) # Frankfurt Airport (longitude, latitude)
polygon_at_distance(fra, 40)
}
}
\seealso{
Other spatial: 
\code{\link{bbox_at_distance}()},
\code{\link{bounding_box}()},
\code{\link{crs_tc}()},
\code{\link{filter_positions_at_range}()},
\code{\link{polygon_from_coords}()}
}
\concept{spatial}
