% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr.R
\name{read_cpr}
\alias{read_cpr}
\title{Read Network Manager's archived CPR file}
\usage{
read_cpr(file, delim = ";")
}
\arguments{
\item{file}{A (gzipped) file containing archived CPR's as received
received and processed by the Network Manager.
Filenames are something like
\code{1.201702061001tacop304ARCHIVED_OPLOG_ALL_CPR.gz}}

\item{delim}{the character delimiting the fields [default = ";"]}
}
\value{
A dataframe of correlated position reports (CPR's)
}
\description{
The payload of an archived CPR file contains the following fields:
\tabular{lll}{
  \strong{Name}       \tab \strong{Description} \tab \strong{Type} \cr
  cpm_id              \tab CPR Message ID \tab int \cr
  tact_id             \tab ETFMS flight ID \tab int \cr
  timestamp_etfms     \tab ETFMS reception timestamp \tab char (YY/MM/DD hh:mm:ss) \cr
  timestamp_track     \tab track timestamp \tab char (YY/MM/DD hh:mm:ss) \cr
  block               \tab Block number \tab int \cr
  record              \tab Record number (index within a block) \tab int \cr
  entry_node_sac      \tab Entry Node system area code \tab int \cr
  entry_node_sic      \tab Entry Node system identifier code \tab int \cr
  callsign            \tab Callsign of flight \tab char \cr
  adep_icao           \tab ICAO code of Departure aerodrome \tab char \cr
  ades_icao           \tab ICAO code of Destination aerodrome \tab char \cr
  eobt                \tab Estimated Off-block Time \tab char (YY/MM/DD hh:mm:ss) \cr
  longitude           \tab Longitude (WGS84) \tab decimal degrees  \cr
  latitude            \tab  Latitude (WGS84) \tab  decimal degrees \cr
  flight_level        \tab  Flight level \tab  integer \cr
  track_service       \tab Track service \tab char (Begin, Continuing, End or Begin_And_End) \cr
  ssr_code            \tab SSR code \tab oct \cr
  track_speed         \tab Track velocity (in Knots) \tab int \cr
  track_heading       \tab Track heading \tab char \cr
  climb_rate          \tab Climb or descend rate \tab int \cr
  track_vertical_mode \tab vertical attitude \tab char (Climb, Descend, Level_Flight or Undetermined) \cr
  ifps_id             \tab IFPS flight plan id \tab char \cr
  aircraft_address    \tab ICAO 24-bit address \tab char
}
}
\examples{
\dontrun{
# read CPR's of the 20 sample flights for 5th Feb 2017
cprs_file <- system.file("extdata",
                        "1.201702051001tacop304ARCHIVED_OPLOG_ALL_CPR.gz",
                        package = "trrrj")
cprs05 <- read_cpr(cprs_file)
}
}
\seealso{
Other read/export: 
\code{\link{export_allft_so6}()},
\code{\link{export_apds}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flight_info}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_hourly_adsb}()},
\code{\link{export_model_trajectory}()},
\code{\link{export_movements}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{fix_flights_csvs}()},
\code{\link{fix_nulls_in_callsign}()},
\code{\link{generate_so6}()},
\code{\link{import_flights_csvs}()},
\code{\link{parse_airspace_prisme}()},
\code{\link{read_flights_fr24}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export}
