% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{read_flights_fr24}
\alias{read_flights_fr24}
\title{Read FR24's archived flights information for a day.}
\usage{
read_flights_fr24(path)
}
\arguments{
\item{path}{the filename of the \code{<yyyymmdd>_flights<eventually>.csv} file.
The \code{<yyyymmdd>} part of the \code{path}'s \code{basename}
is extracted into the \code{date} field.}
}
\value{
a data frame of the archived flights that have flown on \code{<yyyymmdd>}.
  Each flight has the following attributes (all optional, i.e. eventually empty,
  but  \code{flight_id}):
  \itemize{
    \item \code{date}         date of the flight
    \item \code{flight_id}    unique decimal identifier for the flight leg
    \item \code{aircraft_id}  mode-S address in hexadecimal
    \item \code{reg unique}   aircraft text identifier
    \item \code{equip}        equipment, e.g. \code{B733}
    \item \code{callsign}     callsign, e.g. \code{NAX11S}
    \item \code{flight}       flight number, e.g. \code{DY311}
    \item \code{schd_from}    IATA code for scheduled departure airport, e.g. \code{KKN}
    \item \code{schd_to}      IATA code for scheduled arrival airport
    \item \code{real_to}      IATA code for actual arrival airport (when diverted)
  }
}
\description{
\code{read_flights_fr24} returns all flights as archived by FR24
for a given day (see \code{path} parameter for how the date is
encoded).
}
\details{
The flights listed are the ones known to FR24 and fitting the
the bounding box as contractually agreed.
}
\examples{
\dontrun{
# flights, as recorded by FR24, on 20170205 (5th Feb 2017)
flt_fr24_file <- system.file("extdata",
                         "20170205_flights.csv",
                         package = "trrrj")
read_flights_fr24(flt_fr24_file)
}

}
\seealso{
Other read/export functions: 
\code{\link{export_allft_so6}()},
\code{\link{export_event_so6}()},
\code{\link{export_event_trajectory}()},
\code{\link{export_flights_at_airport_fr24}()},
\code{\link{export_flights_fr24}()},
\code{\link{export_positions_at_airport_fr24}()},
\code{\link{export_positions_fr24}()},
\code{\link{import_flights_csvs}()},
\code{\link{read_cpr}()},
\code{\link{read_positions_fr24}()},
\code{\link{read_so6}()}
}
\concept{read/export functions}
