% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr24.R
\name{read_positions_fr24}
\alias{read_positions_fr24}
\title{Read FR24's archived positions information for a flight.}
\usage{
read_positions_fr24(path)
}
\arguments{
\item{path}{the filename of the \code{<yyyymmdd>_<flight_id>_positions.csv} file}
}
\value{
a data frame of the archived flight's positions for date \code{<yyyymmdd>}.
  Each position report has the following attributes (all mandatory, i.e. non empty,
  but \code{speed}, \code{squawk} and \code{vert_speed}):
  \itemize{
    \item \code{flight_id}   FR24 unique (within a day?) flight leg identifier
    \item \code{timestamp}   \code{snapshot_id} converted to a date/time
    \item \code{snapshot_id} time of sample in Unix epoch seconds,
          i.e. since 1st Jan 1970 00:00:00 UTC
    \item \code{altitude}    altitude (feet). Barometric relative to ISO 1013 hPa
          pressure in flight, always 0 on ground.
    \item \code{heading}     0-359 degrees
    \item \code{latitude}    latitude in decimal degrees
    \item \code{longitude}   longitude in decimal degrees
    \item \code{radar_id}    unique identifier of primary ground receiver
    \item \code{speed}       ground speed (knots)
    \item \code{squawk}      code broadcast by airplane, represented as four
          octal digits
    \item \code{vert_speed}  vertical speed (feet/min). Calculated from rate
          of altitude changes, 0 for level flight or empty if unavailable
    \item \code{on_ground}   1 for on ground, 0 for airborne.
          Calculated from low altitude and ground speed
  }
}
\description{
\code{read_positions_fr24} returns all position reports as archived by FR24
for a given \code{flight_id} flown on a specific day (see \code{path}
parameter for how the date and flight ID are encoded).
}
\details{
The positions listed are the ones known to FR24 and fitting the
the bounding box as contractually agreed.
}
\examples{
\dontrun{
# positions, as recorded by FR24, for flight 207535493 on 20170205 (5th Feb 2017)
pos_fr24_file <- system.file("extdata",
                             "20170205_positions",
                             "20170205_207535493.csv",
                             package = "trrrj")
read_positions_fr24(pos_fr24_file)
}
}
\seealso{
Other read/export functions: \code{\link{export_allft_so6}},
  \code{\link{export_event_so6}},
  \code{\link{export_event_trajectory}},
  \code{\link{export_flights_at_airport_fr24}},
  \code{\link{export_flights_fr24}},
  \code{\link{export_positions_at_airport_fr24}},
  \code{\link{export_positions_fr24}},
  \code{\link{import_flights_csvs}},
  \code{\link{read_cpr}}, \code{\link{read_flights_fr24}},
  \code{\link{read_so6}}
}
\concept{read/export functions}
