% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_positions}
\alias{smooth_positions}
\title{Smooth longitude, latitude and altitude.}
\usage{
smooth_positions(df, n = 8, type = "t")
}
\arguments{
\item{df}{a data frame of trajectory positions}

\item{n}{from pracma::movavg}

\item{type}{from pracma::movavg}
}
\value{
return a smoothed trajectory
}
\description{
Smooth longitude, latitude and altitude.
}
\examples{
smooth_positions(egll_positions)

}
