% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_positions}
\alias{smooth_positions}
\title{Smooth longitude, latitude and altitude.}
\usage{
smooth_positions(df, n = 5, type = "t")
}
\arguments{
\item{df}{a data frame of trajectory positions}

\item{n}{from pracma::movavg}

\item{type}{from pracma::movavg}
}
\value{
return a smoothed trajectory
}
\description{
Smooth longitude, latitude and altitude.
}
\examples{
data(list = c("egll_positions_fr24"), package = "trrrj")
smooth_positions(egll_positions_fr24)

}
\seealso{
Other analysis: 
\code{\link{cumulative_distance}()},
\code{\link{cumulative_time}()},
\code{\link{extract_segment}()},
\code{\link{filter_outlier}()},
\code{\link{resample}()}
}
\concept{analysis}
